Added randomized heights

+++ VS2017/src/Heights.h
@@ -7,14 +7,14 @@
 #include <numeric>
 #include <vector>
 
-float map(float input, int start, int end, int newStart, int newEnd)
+float map(float input, float start, float end, float newStart, float newEnd)
 {
 	float output = newStart + ((newEnd - newStart) / (end - start)) * (input - start);
 	return output;
 }
 
 
-float randomizeHeight(int startRange, int endRange)
+float randomizeHeight(float startRange, float endRange)
 {
 	float result = static_cast <float> (rand()) / static_cast <float> (RAND_MAX);
 	return map(result, 0, 1, startRange, endRange);
